+++ data/schema.graphql
@@ -43,6 +43,9 @@ type PassAddedPayload {
 }
 
 input PassChangePasswordInput {
+  """password global id"""
+  id: String!
+
   """pass new password"""
   password: String!
   clientMutationId: String
@@ -84,7 +87,6 @@ input PassCreateInput {
   website: String!
   login: String!
   password: String!
-  user: String!
   clientMutationId: String
 }
