+++ src/App.tsx
@@ -4,7 +4,7 @@ import {createStackNavigator} from 'react-navigation-stack';
 import LoginScreen from './screens/Login';
 import HomeScreen from './screens/Home';
 import DetailScreen from './relay/PassDetail';
-import AddPassScreen from './screens/AddPass';
+import PassFormScreen from './screens/PassForm';
 import SignUpScreen from './screens/SignUp';
 
 const AppNavigator = createStackNavigator(
@@ -12,7 +12,7 @@ const AppNavigator = createStackNavigator(
     Login: {screen: LoginScreen},
     Home: {screen: HomeScreen},
     Detail: {screen: DetailScreen},
-    Add: {screen: AddPassScreen},
+    PassForm: {screen: PassFormScreen},
     SignUp: {screen: SignUpScreen},
   },
   {initialRouteName: 'Login'},
